requirements.txt
discord.py
simpleaudio
python-dotenv
